Rename *MapStorage to *CrudOperations
