+++ model/map/src/main/java/org/keycloak/models/map/storage/chm/ConcurrentHashMapStorageProviderFactory.java
@@ -101,7 +101,7 @@ public class ConcurrentHashMapStorageProviderFactory implements AmphibianProvide
 
     private static final Logger LOG = Logger.getLogger(ConcurrentHashMapStorageProviderFactory.class);
 
-    private final ConcurrentHashMap<String, ConcurrentHashMapStorage<?,?,?>> storages = new ConcurrentHashMap<>();
+    private final ConcurrentHashMap<String, ConcurrentHashMapCrudOperations<?,?,?>> storages = new ConcurrentHashMap<>();
 
     private final Map<String, StringKeyConverter> keyConverters = new HashMap<>();
 
@@ -217,7 +217,7 @@ public class ConcurrentHashMapStorageProviderFactory implements AmphibianProvide
     }
 
     @SuppressWarnings("unchecked")
-    private void storeMap(String mapName, ConcurrentHashMapStorage<?, ?, ?> store) {
+    private void storeMap(String mapName, ConcurrentHashMapCrudOperations<?, ?, ?> store) {
         if (mapName != null) {
             File f = getFile(mapName);
             try {
@@ -235,22 +235,22 @@ public class ConcurrentHashMapStorageProviderFactory implements AmphibianProvide
     }
 
     @SuppressWarnings("unchecked")
-    private <K, V extends AbstractEntity & UpdatableEntity, M> ConcurrentHashMapStorage<K, V, M> loadMap(String mapName,
-      Class<M> modelType, EnumSet<Flag> flags) {
+    private <K, V extends AbstractEntity & UpdatableEntity, M> ConcurrentHashMapCrudOperations<K, V, M> loadMap(String mapName,
+                                                                                                                Class<M> modelType, EnumSet<Flag> flags) {
         final StringKeyConverter kc = keyConverters.getOrDefault(mapName, defaultKeyConverter);
         Class<?> valueType = ModelEntityUtil.getEntityType(modelType);
         LOG.debugf("Initializing new map storage: %s", mapName);
 
-        ConcurrentHashMapStorage<K, V, M> store;
+        ConcurrentHashMapCrudOperations<K, V, M> store;
         if(modelType == SingleUseObjectValueModel.class) {
-            store = new SingleUseObjectConcurrentHashMapStorage(kc, CLONER) {
+            store = new SingleUseObjectConcurrentHashMapCrudOperations(kc, CLONER) {
                 @Override
                 public String toString() {
                     return "ConcurrentHashMapStorage(" + mapName + suffix + ")";
                 }
             };
         } else {
-            store = new ConcurrentHashMapStorage(modelType, kc, CLONER) {
+            store = new ConcurrentHashMapCrudOperations(modelType, kc, CLONER) {
                 @Override
                 public String toString() {
                     return "ConcurrentHashMapStorage(" + mapName + suffix + ")";
@@ -286,7 +286,7 @@ public class ConcurrentHashMapStorageProviderFactory implements AmphibianProvide
     }
 
     @SuppressWarnings("unchecked")
-    public <K, V extends AbstractEntity & UpdatableEntity, M> ConcurrentHashMapCrudOperations<V, M> getStorage(
+    public <K, V extends AbstractEntity & UpdatableEntity, M> CrudOperations<V, M> getStorage(
       Class<M> modelType, Flag... flags) {
         EnumSet<Flag> f = flags == null || flags.length == 0 ? EnumSet.noneOf(Flag.class) : EnumSet.of(flags[0], flags);
         String name = getModelName(modelType, modelType.getSimpleName());
@@ -295,7 +295,7 @@ public class ConcurrentHashMapStorageProviderFactory implements AmphibianProvide
          *   "... the computation [...] must not attempt to update any other mappings of this map."
          */
 
-        return (ConcurrentHashMapCrudOperations<V, M>) storages.computeIfAbsent(name, n -> loadMap(name, modelType, f));
+        return (CrudOperations<V, M>) storages.computeIfAbsent(name, n -> loadMap(name, modelType, f));
     }
 
     public StringKeyConverter<?> getKeyConverter(Class<?> modelType) {
